refactor: DailyBoxOfficeItem파일 BoxOffice파일에 병합, MovieInfomaion파일 Movie파일에 병합
